bibcujae/requirements.txt
barcode_generator==0.1rc15
Django==4.1.7
djangorestframework==3.14.0
openpyxl==3.1.2
Pillow==9.5.0
rest_framework_simplejwt==0.0.2
